data/functions/bq-to-datastore/requirements.txt
google-cloud-bigquery==1.20.0
google-cloud-datastore==1.9.0
